WIP Split up DsForeign/TcForeign
